due to aspectj not allows for extending, so using abstract class to do it
